using Spring Boot's auto configuration to create an ObjectMapper. ObjectMapperProvider is used by the client and tests.
