+++ cassandra-persistence/build.gradle
@@ -1,3 +1,16 @@
+/*
+ *  Copyright 2021 Netflix, Inc.
+ *  <p>
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ *  the License. You may obtain a copy of the License at
+ *  <p>
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  <p>
+ *  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ *  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ */
+
 dependencies {
     implementation project(':conductor-common')
     implementation project(':conductor-core')
@@ -11,5 +24,7 @@ dependencies {
     testImplementation("org.cassandraunit:cassandra-unit:${revCassandraUnit}") {
         exclude group: "com.datastax.cassandra", module: "cassandra-driver-core"
     }
+
     testImplementation project(':conductor-core').sourceSets.test.output
+    testImplementation project(':conductor-common').sourceSets.test.output
 }
