+++ cassandra-persistence/src/test/java/com/netflix/conductor/cassandra/dao/CassandraDAOTest.java
@@ -19,7 +19,7 @@ import com.netflix.conductor.cassandra.config.CassandraProperties;
 import com.netflix.conductor.cassandra.dao.CassandraBaseDAO.WorkflowMetadata;
 import com.netflix.conductor.cassandra.util.EmbeddedCassandra;
 import com.netflix.conductor.cassandra.util.Statements;
-import com.netflix.conductor.common.config.ObjectMapperConfiguration;
+import com.netflix.conductor.common.config.TestObjectMapperConfiguration;
 import com.netflix.conductor.common.metadata.events.EventExecution;
 import com.netflix.conductor.common.metadata.events.EventHandler;
 import com.netflix.conductor.common.metadata.tasks.Task;
@@ -60,7 +60,7 @@ import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-@ContextConfiguration(classes = {ObjectMapperConfiguration.class})
+@ContextConfiguration(classes = {TestObjectMapperConfiguration.class})
 @RunWith(SpringRunner.class)
 public class CassandraDAOTest {
