+++ common/src/main/java/com/netflix/conductor/common/config/ObjectMapperBuilderConfiguration.java
@@ -1,0 +1,36 @@
+/*
+ *  Copyright 2021 Netflix, Inc.
+ *  <p>
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ *  the License. You may obtain a copy of the License at
+ *  <p>
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  <p>
+ *  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ *  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ */
+
+package com.netflix.conductor.common.config;
+
+import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_IGNORED_PROPERTIES;
+import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_NULL_FOR_PRIMITIVES;
+import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES;
+
+@Configuration
+public class ObjectMapperBuilderConfiguration {
+
+    /**
+     * Disable features like {@link ObjectMapperProvider#getObjectMapper()}.
+     */
+    @Bean
+    public Jackson2ObjectMapperBuilderCustomizer conductorJackson2ObjectMapperBuilderCustomizer() {
+        return builder -> builder.featuresToDisable(FAIL_ON_UNKNOWN_PROPERTIES,
+                FAIL_ON_IGNORED_PROPERTIES,
+                FAIL_ON_NULL_FOR_PRIMITIVES);
+    }
+}
