+++ server/build.gradle
@@ -1,3 +1,16 @@
+/*
+ *  Copyright 2021 Netflix, Inc.
+ *  <p>
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ *  the License. You may obtain a copy of the License at
+ *  <p>
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  <p>
+ *  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ *  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ */
+
 plugins {
     id 'application'
     id 'org.springframework.boot'
@@ -27,6 +40,13 @@ dependencies {
     implementation "org.springdoc:springdoc-openapi-ui:${revOpenapi}"
 
     runtimeOnly 'com.netflix.spectator:spectator-reg-micrometer'
+
+    testImplementation project(':conductor-rest')
+    testImplementation project(':conductor-common')
+    testImplementation "io.grpc:grpc-testing:${revGrpc}"
+    testImplementation "com.google.protobuf:protobuf-java:${revProtoBuf}"
+    testImplementation "io.grpc:grpc-protobuf:${revGrpc}"
+    testImplementation "io.grpc:grpc-stub:${revGrpc}"
 }
 
 jar {
