+++ server/src/test/java/com/netflix/conductor/common/config/ConductorObjectMapperTest.java
@@ -1,15 +1,16 @@
 /*
- * Copyright 2020 Netflix, Inc.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
+ *  Copyright 2021 Netflix, Inc.
+ *  <p>
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ *  the License. You may obtain a copy of the License at
+ *  <p>
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  <p>
+ *  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ *  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
  */
+
 package com.netflix.conductor.common.config;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -22,7 +23,7 @@ import com.netflix.conductor.common.run.Workflow;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.io.IOException;
@@ -35,9 +36,13 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-@ContextConfiguration(classes = {ObjectMapperConfiguration.class})
+/**
+ * Tests the customized {@link ObjectMapper} that is used by
+ * {@link com.netflix.conductor.Conductor} application.
+ */
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @RunWith(SpringRunner.class)
-public class ObjectMapperTest {
+public class ConductorObjectMapperTest {
 
     @Autowired
     ObjectMapper objectMapper;
@@ -47,7 +52,7 @@ public class ObjectMapperTest {
         assertTrue(objectMapper.canSerialize(Any.class));
 
         Struct struct1 = Struct.newBuilder().putFields(
-            "some-key", Value.newBuilder().setStringValue("some-value").build()
+                "some-key", Value.newBuilder().setStringValue("some-value").build()
         ).build();
 
         Any source = Any.pack(struct1);
@@ -61,8 +66,8 @@ public class ObjectMapperTest {
         Struct struct2 = dest.unpack(Struct.class);
         assertTrue(struct2.containsFields("some-key"));
         assertEquals(
-            struct1.getFieldsOrThrow("some-key").getStringValue(),
-            struct2.getFieldsOrThrow("some-key").getStringValue()
+                struct1.getFieldsOrThrow("some-key").getStringValue(),
+                struct2.getFieldsOrThrow("some-key").getStringValue()
         );
     }
